Finalises version 1.0 of search using Algolia.
Refactored the search component into multiple smaller components. Prevents search results from loading until there is a search query. Lots of styling changes around the search itself.
