+++ src/components/search/search.scss
@@ -1,47 +1,65 @@
 @import "../../assets/styles/main";
 
 .ais-SearchBox-input {
-    width: 100%;
-    padding: 1em;
+  width: 100%;
+  padding: 1em;
+  margin-top: 0.5em;
 }
 
 .ais-InfiniteHits-list {
-    max-height: 30em;
-    overflow-y: auto;
+  overflow-y: auto;
+  padding: 1em 0;
 }
 
 .ais-InfiniteHits-item {
-    padding: 0.25em;
+  padding: 0.25em;
+}
+
+.ais-InfiniteHits-loadMore {
+  font-size: 100%;
+  font-family: inherit;
+  border-width: 1px;
+  cursor: pointer;
+  -ms-flex-pack: center;
+  justify-content: center;
+  padding: calc(0.5em - 1px) 1em;
+  text-align: center;
+  white-space: nowrap;
+  background-color: $primary-color;
+  border-color: transparent;
+  color: #fff;
+  display: block;
+  margin: 0 auto;
 }
 
 .navbar-item.searchButton {
-    margin-left: 1em;
-    color: inherit;
-    font-size: 100%;
-    font-family: inherit;
-    border: 0;
-    padding: 0.5em;
-    background: none;
-    cursor: pointer;
+  margin-left: 1em;
+  color: inherit;
+  font-size: 100%;
+  font-family: inherit;
+  border: 0;
+  padding: 0.5em;
+  background: none;
+  cursor: pointer;
 }
 
 .navbar-item.searchButton:hover {
-    color: $primary-color;
+  color: $primary-color;
 }
 
 .searchIcon {
-    padding-right: 1em;
+  padding-right: 1em;
 }
 
 .searchCloseButton {
-    font-size: 100%;
-    font-family: inherit;
-    border: 0;
-    padding: 0;
-    background: none;
-    color: $primary-color;
-    position: absolute;
-    top: 0;
-    right: 0;
-    cursor: pointer;
+  font-size: 100%;
+  font-family: inherit;
+  border: 0;
+  padding: 0;
+  background: none;
+  color: $primary-color;
+  position: absolute;
+  top: 0;
+  right: 0;
+  cursor: pointer;
 }
